requirements.txt
bert_score
evaluate~=0.4.5
optimum~=1.26.1
pydantic~=2.11.7
PyYAML~=6.0.2
torch~=2.7.1
tqdm
transformers~=4.53.3

bert-score~=0.3.13
scikit-learn~=1.7.1
datasets~=4.0.0
numpy~=2.3.1